Consolidate action param handler into existing WooCommerce_Payments note class
